+++ src/Hacks/Universal/DecemberSnow.cpp
@@ -30,11 +30,23 @@ class $modify (MenuLayer)
 
         std::time_t currentTime = std::time(nullptr);
         std::tm* localTime = std::localtime(&currentTime);
+        auto mon = localTime->tm_mon + 1;
 
-        if (localTime->tm_mon + 1 == 12)
+        if (mon == 12 || mon == 1)
         {
             auto snow = CCParticleSnow::create();
             snow->setID("snow"_spr);
+
+            static bool first = false;
+
+            if (!first)
+            {
+                first = true;
+            }
+            else
+            {
+                snow->update(15);
+            }
 
             this->addChild(snow, 420);
         }